IN-249 refactor: mobile version
