+++ src/shared/ui/Pagination/Pagination.module.scss
@@ -81,7 +81,7 @@
   }
 }
 
-@media screen and (width <= 640px) {
+@media screen and (width <= 920px) {
   .container {
     flex-direction: column-reverse;
     justify-content: center;
