--- src/widgets/profile/ui/MyPayments/MobileMyPayments/MobileMyPayments.module.scss
@@ -1,17 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  margin-bottom: 20px;
-}
-
-.card {
-  padding: 19px 12px 17px;
-  background-color: var(--dark-500);
-  border: 1px solid var(--dark-300);
-}
-
-.row {
-  display: flex;
-  justify-content: space-between;
-}
